+++ Src/Log.h
@@ -19,7 +19,7 @@
 
 namespace RabbitCommon {
 
-#ifdef HAVE_GUI
+#ifdef HAVE_RABBITCOMMON_GUI
     void OpenLogConfigureFile();
     void OpenLogFile();
     void OpenLogFolder();