+++ Tests/MainWindow.cpp
@@ -6,22 +6,20 @@
 #include "ui_MainWindow.h"
 
 #ifdef HAVE_ABOUT
-    #include "DlgAbout/DlgAbout.h"
+    #include "DlgAbout.h"
 #endif
 #ifdef HAVE_UPDATE
-    #include "FrmUpdater/FrmUpdater.h"
-#endif
-#ifdef HAVE_ADMINAUTHORISER
-    #include "AdminAuthoriser/adminauthoriser.h"
+    #include "FrmUpdater.h"
 #endif
 #ifdef BUILD_QUIWidget
     #include "QUIWidget/QUIWidget.h"
 #endif
 #include "RabbitCommonTools.h"
 #include "RabbitCommonEncrypt.h"
-#include "Log.h"
-#include "FolderBrowser/FolderBrowser.h"
-#include "Style/FrmStyle.h"
+#ifdef HAVE_RABBITCOMMON_GUI
+#include "FrmStyle.h"
+#include "DockFolderBrowser.h"
+#endif
 
 #include <QDir>
 #include <QDebug>
@@ -37,12 +35,14 @@ MainWindow::MainWindow(QWidget *parent) :
 {
     ui->setupUi(this);
 
-    // [Use CFolderBrowser]
-    CFolderBrowser* pDock = new CFolderBrowser(tr("Folder browser"), this);
+#ifdef HAVE_RABBITCOMMON_GUI
+    // [Use CDockFolderBrowser]
+    CDockFolderBrowser* pDock = new CDockFolderBrowser(tr("Folder browser"), this);
     addDockWidget(Qt::LeftDockWidgetArea, pDock);
     // Add the action of dock to menu
     ui->menuTools->addAction(pDock->toggleViewAction());
-    // [Use CFolderBrowser]
+    // [Use CDockFolderBrowser]
+#endif
 }
 
 MainWindow::~MainWindow()
@@ -70,6 +70,7 @@ void MainWindow::on_actionAbout_triggered()
 
 void MainWindow::on_actionUpdate_triggered()
 {
+#ifdef HAVE_UPDATE
     // [Use CFrmUpdater]
     CFrmUpdater* m_pfrmUpdater = new CFrmUpdater();
     QIcon icon = windowIcon();
@@ -85,7 +86,7 @@ void MainWindow::on_actionUpdate_triggered()
         m_pfrmUpdater->show();
     #endif
     // [Use CFrmUpdater]
-
+#endif
 #ifdef BUILD_QUIWidget
     QUIWidget::setFormInCenter(m_pfrmUpdater);
 #endif
@@ -93,7 +94,6 @@ void MainWindow::on_actionUpdate_triggered()
 
 void MainWindow::on_pushButton_clicked()
 {
-#ifdef HAVE_ADMINAUTHORISER
 #if defined(Q_OS_LINUX) && !defined(Q_OS_ANDROID)
     QString szCmd = "mkdir";
     QStringList paras;
@@ -104,12 +104,11 @@ void MainWindow::on_pushButton_clicked()
 #elif defined(Q_OS_WINDOWS)
     RabbitCommon::CTools::executeByRoot("regedit", QStringList());
 #endif
-#endif
 }
 
 void MainWindow::on_actionStype_triggered()
 {
-#ifdef HAVE_GUI
+#ifdef HAVE_RABBITCOMMON_GUI
     CFrmStyle* s = new CFrmStyle();
     s->show();
 #endif
@@ -135,17 +134,23 @@ void MainWindow::on_pbEncrypt_clicked()
 
 void MainWindow::on_actionOpen_log_configure_triggered()
 {
+#ifdef HAVE_RABBITCOMMON_GUI
     RabbitCommon::CTools::Instance()->OpenLogConfigureFile();
+#endif
 }
 
 void MainWindow::on_actionOpen_log_file_triggered()
 {
+#ifdef HAVE_RABBITCOMMON_GUI
     RabbitCommon::CTools::Instance()->OpenLogFile();
+#endif
 }
 
 void MainWindow::on_actionOpen_log_folder_triggered()
 {
+#ifdef HAVE_RABBITCOMMON_GUI
     RabbitCommon::CTools::Instance()->OpenLogFolder();
+#endif
 }
 
 void MainWindow::on_pbGenerateCoreFile_clicked()